requirements.txt
python-binance==1.0.35
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
APScheduler>=3.10.4
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
aiohttp>=3.9.0
pytz>=2024.1
pytest>=8.0.0
pytest-mock>=3.12.0
